+++ rocketgram/api/input_venue_message_content.py
@@ -22,3 +22,5 @@ class InputVenueMessageContent(InputMessageContent):
     address: str
     foursquare_id: Optional[str] = None
     foursquare_type: Optional[str] = None
+    google_place_id: Optional[str] = None
+    google_place_type: Optional[str] = None